Split API v1 routes and add task/sentence detail endpoints
